pyproject.toml
[build-system]
requires = ["setuptools>=67.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "speech-commands"
version = "0.1.0"
description = "Speech Commands Classification with Transformers"
authors = [
    { name = "Jakub Kępka" },
    { name = "Damian Kąkol" },
]
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
    "transformers>=4.38.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pyyaml>=6.0.1",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-cov>=5.0.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
